Better align book details when shown in a grid
In a book grid, the title, author and status are shown in a column
with no real layout constraints. This means if a book has a long
title, its author and status won't line up with a different book
with a short title, which makes scanning across the grid annoying.

To fix this, give the title the flex-grow property so it'll expand
to fill the available room. Authors and status will then be
bottom-aligned.