feat: optimize Dockerfile structure
Most of the time & size of building the docker image comes from all of
the Ubuntu applications and downloads that have to be made.

By loading the application after downloading the operating system
packages, it makes the docker image more efficient both in terms of the
size of the layers and the time it takes to build the docker image. This
greatly eases rapid iteration on the docker image.