feat: add full unit tests for CalibreMetadataProvider
After adding errors for MetadataDto, it seemed like a good idea to add
proper unit testing that leverages the logic in the fetchMetadata()
function. So we made extensive use of the mockk library which was used
elsewhere in the codebase to mock out the ProcessBuilder and then test
to ensure that we only get errors when there is a problem fetching the
metadata.