Update DatabazeKnihMetadataProvider.kt
I was looking for another way to fetch ISBN, because the page count and language was also missing.

This version should fetch a second details page (the "more info" AJAX endpoint) and parses the ISBN from there, not from the initial book page.

I tried to build image from source to test it, but I couldn't get frontend to run. Is there some easy way to test it? Or can you please look at this code and test it?